refactor branding page and extract out to new BrandingCore component
